Fix bugs
Tell user in the caption if the file is saved. Fix ctrl backspacing when
there are newline characters. Add just_typed=True in more places so that
the program knows when text has justed been typed, or the screen has
been moved, so that it can readjust the scroll_top and mark the file as
unsaved.